+++ mail_private/static/src/js/test_private.js
@@ -1,17 +1,30 @@
-/*  Copyright 2018 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
+/*  Copyright 2018-2019 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
+    Copyright 2019 Artem Rafailov <https://it-projects.info/team/Ommo73/>
     License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).*/
-odoo.define('mail_private.tour', function (require) {
-    "use strict";
+odoo.define('mail_private.tour', function(require) {
+"use strict";
 
-    var tour = require("web_tour.tour");
     var core = require('web.core');
+    var tour = require('web_tour.tour');
     var _t = core._t;
 
     var email = 'mail_private test email';
-    var steps = [{
-            trigger: '.o_thread_message strong.o_mail_redirect:contains("Agrolait")',
-            content: _t("Open Partners Form"),
+    var steps = [tour.STEPS.SHOW_APPS_MENU_ITEM, {
+            trigger: '.fa.fa-cog.o_mail_channel_settings',
+            content: _t('Select channel settings'),
             position: 'bottom',
+        }, {
+            trigger: '.nav-link:contains("Members")',
+            content: _t('Go to the list of subscribers'),
+            position: 'bottom',
+        }, {
+            trigger: '.o_data_cell:contains("YourCompany, Marc Demo")',
+            content: _t("Select a user"),
+            position: "bottom",
+        }, {
+            trigger: '.o_form_uri.o_field_widget:contains("YourCompany, Marc Demo")',
+            content: _t("Go to user page"),
+            position: "bottom"
         }, {
             trigger: "button.oe_compose_post_private",
             content: _t("Click on Private mail creating button"),
@@ -26,7 +39,7 @@ odoo.define('mail_private.tour', function (require) {
             content: _t("Uncheck all Followers"),
             timeout: 10000,
         }, {
-            trigger: "div.o_composer_suggested_partners input:first",
+            trigger: "div.o_composer_suggested_partners",
             content: _t("Check the first one"),
         }, {
             trigger: "textarea.o_composer_text_field:first",