+++ mail_private/tests/test_js.py
@@ -1,9 +1,8 @@
-# -*- coding: utf-8 -*-
 # Copyright 2018 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
+# Copyright 2019 Artem Rafailov <https://it-projects.info/team/Ommo73/>
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 
 import odoo.tests
-from odoo.api import Environment
 
 
 @odoo.tests.common.at_install(True)
@@ -16,15 +15,8 @@ class TestUi(odoo.tests.HttpCase):
         # that are returned by the backend in module_boot. Without
         # this you end up with js, css but no qweb.
         cr = self.registry.cursor()
-        env = Environment(cr, self.uid, {})
-        env['ir.module.module'].search([('name', '=', 'mail_private')], limit=1).state = 'installed'
-        cr.release()
-
-        env = Environment(self.registry.test_cr, self.uid, {})
-        partners = env['res.partner'].search([])
-        new_follower = env['res.partner'].search([('name', 'ilike', 'Ja')])
-        for partner in partners:
-            partner.message_subscribe(new_follower.ids, [])
+        self.env['ir.module.module'].search([('name', '=', 'mail_private')], limit=1).state = 'installed'
+        cr._lock.release()
 
         self.phantom_js("/web",
                         "odoo.__DEBUG__.services['web_tour.tour'].run('mail_private_tour', 1000)",